+++ views/admin-cms-content-element/index.php
@@ -20,26 +20,66 @@ if ($content_id = \Yii::$app->request->get('content_id'))
      */
     $cmsContent = \skeeks\cms\models\CmsContent::findOne($content_id);
 }
-
-$filter = new \yii\base\DynamicModel([
-    'query',
-]);
-
-$filter->addRule('query', 'string');
-
 ?>
 <? $pjax = \yii\widgets\Pjax::begin(); ?>
 
-    <? $form = \skeeks\cms\base\widgets\ActiveForm::begin([
-        'method' => 'get',
-        //'usePjax' => false,
-        'enableAjaxValidation' => false,
-        /*'options' =>
-        [
-            'data-pjax' => true
-        ]*/
+    <? $form = \skeeks\cms\modules\admin\widgets\filters\AdminFiltersForm::begin([
+        'action' => '/' . \Yii::$app->request->pathInfo,
+        'namespace' => \Yii::$app->controller->uniqueId . "_" . $content_id
     ]); ?>
-        <?/*= $form->field($filter, 'query'); */?>
+
+        <?= \yii\helpers\Html::hiddenInput('content_id', $content_id) ?>
+
+        <?= $form->field($searchModel, 'id'); ?>
+        <?= $form->field($searchModel, 'name')->setVisible(); ?>
+
+        <?= $form->field($searchModel, 'active')->listBox(\yii\helpers\ArrayHelper::merge([
+            '' => ' - '
+        ], \Yii::$app->cms->booleanFormat()), [
+            'size' => 1
+        ]); ?>
+
+        <?= $form->field($searchModel, 'section')->listBox(\yii\helpers\ArrayHelper::merge([
+            '' => ' - '
+        ], \skeeks\cms\helpers\TreeOptions::getAllMultiOptions()),
+        [
+            'unselect' => ' - ',
+            'size' => 1
+        ]); ?>
+
+
+        <?= $form->field($searchModel, 'has_image')->checkbox(\Yii::$app->formatter->booleanFormat, false); ?>
+        <?= $form->field($searchModel, 'has_full_image')->checkbox(\Yii::$app->formatter->booleanFormat, false); ?>
+
+
+        <?= $form->field($searchModel, 'created_by')->widget(\skeeks\cms\modules\admin\widgets\formInputs\SelectModelDialogUserInput::className()); ?>
+        <?= $form->field($searchModel, 'updated_by')->widget(\skeeks\cms\modules\admin\widgets\formInputs\SelectModelDialogUserInput::className()); ?>
+
+
+        <?= $form->field($searchModel, 'created_at_from')->widget(
+            \kartik\datetime\DateTimePicker::className()
+        ); ?>
+        <?= $form->field($searchModel, 'created_at_to')->widget(
+            \kartik\datetime\DateTimePicker::className()
+        ); ?>
+
+        <?= $form->field($searchModel, 'updated_at_from')->widget(
+            \kartik\datetime\DateTimePicker::className()
+        ); ?>
+        <?= $form->field($searchModel, 'updated_at_to')->widget(
+            \kartik\datetime\DateTimePicker::className()
+        ); ?>
+
+        <?= $form->field($searchModel, 'published_at_from')->widget(
+            \kartik\datetime\DateTimePicker::className()
+        ); ?>
+        <?= $form->field($searchModel, 'published_at_to')->widget(
+            \kartik\datetime\DateTimePicker::className()
+        ); ?>
+
+        <?= $form->field($searchModel, 'code'); ?>
+
+
     <? $form::end(); ?>
 
     <?= \skeeks\cms\modules\admin\widgets\GridViewStandart::widget([
@@ -55,7 +95,7 @@ $filter->addRule('query', 'string');
         'columns' => \skeeks\cms\controllers\AdminCmsContentElementController::getColumns($cmsContent, $dataProvider)
     ]); ?>
 
-<? \yii\widgets\Pjax::end() ?>
+<? \yii\widgets\Pjax::end(); ?>
 
 <? \yii\bootstrap\Alert::begin([
     'options' => [